pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-all-techniques"
version = "0.1.0"
description = "So many kinds of rag techniques!"
readme = "README.md"
requires-python = ">=3.13.2"
license = {file = "LICENSE"}
authors = [
    {name = "Andy", email = "andy@example.com"}
]


dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.0.0",
    "tqdm>=4.65.0",
    "neo4j==5.28.2",
    "spacy==3.8.7",
    "pip>=25.2",
    "jieba>=0.42.1",
    "rank-bm25>=0.2.2",
    "tavily-python>=0.3.0",
    "tiktoken>=0.11.0",
]

[project.optional-dependencies]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["app"]


